claude: simplify the hypothesistooling build-tooling package
Collapse the multi-"project" abstraction to a single project and remove
accumulated indirection in the build tooling:

- remove the projects/ subpackage and all_projects(); rename
  projects/hypothesis.py to release.py and fold releasemanagement.py into it
- prune dead helpers (extract_assignment, update_markdown_changelog,
  parse_version, current_branch, last_committer, changed_files_from_master);
  delete junkdrawer.py, inlining its only live function (once) into installers.py
- move the git/repo helpers into git.py
- drop the package facade: __init__.py is now just the license header and every
  consumer imports directly from the relevant submodule
