+++ tooling/src/hypothesistooling/installers.py
@@ -19,9 +19,20 @@ import shutil
 import subprocess
 
 from hypothesistooling import scripts
-from hypothesistooling.junkdrawer import once
 
 HOME = os.environ["HOME"]
+
+
+def once(fn):
+    def accept():
+        if accept.has_been_called:
+            return
+        fn()
+        accept.has_been_called = True
+
+    accept.has_been_called = False
+    accept.__name__ = fn.__name__
+    return accept
 
 
 def __python_executable(version):
