+++ whole_repo_tests/whole_repo/test_rst_is_valid.py
@@ -12,19 +12,19 @@ import os
 import re
 from pathlib import Path
 
-import hypothesistooling as tools
-from hypothesistooling.projects import hypothesis as hp
+from hypothesistooling.git import all_files
+from hypothesistooling.release import HYPOTHESIS
 from hypothesistooling.scripts import pip_tool
 
 
 def is_sphinx(f):
     f = os.path.abspath(f)
-    return f.startswith(os.path.join(hp.HYPOTHESIS, "docs"))
+    return f.startswith(os.path.join(HYPOTHESIS, "docs"))
 
 
 ALL_RST = [
     p
-    for p in tools.all_files()
+    for p in all_files()
     if p.name not in ["RELEASE.rst", "RELEASE-sample.rst"] and p.suffix == ".rst"
 ]
